+++ packages/antd-management-fast-design-playground/src/components/displayer/RemarkEditor/index.jsx
@@ -213,8 +213,6 @@ class RemarkEditor extends BaseComponent {
 
     const list = [...dataBackup];
 
-    console.log(list);
-
     this.setState({ dataStorage: [...list] });
   };