+++ packages/example-remote/src/pages/Workflow/FlowDisplayDrawer/index.jsx
@@ -221,6 +221,8 @@ class FlowDisplayDrawer extends BaseVerticalFlexDrawer {
       <Flow
         canEdit={false}
         nodeNameKey={fieldDataWorkflowNode.name.name}
+        listInLineKey={fieldDataWorkflowNode.listInLine.name}
+        listOutLineKey={fieldDataWorkflowNode.listOutLine.name}
         listApproverKey={fieldDataWorkflowNode.listApprover.name}
         personnelNameKey={fieldDataWorkflowNodeApprover.userRealName.name}
         personnelNameLabel={fieldDataWorkflowNodeApprover.userRealName.label}