+++ packages/example-remote/src/pages/WorkflowBranchConditionItem/Common/data.js
@@ -1,0 +1,75 @@
+import { formNameCollection } from '../../../customConfig';
+
+export const fieldData = {
+  ...formNameCollection,
+  workflowBranchConditionItemId: {
+    label: '数据标识',
+    name: 'workflowBranchConditionItemId',
+    helper: '',
+  },
+  workflowId: {
+    label: '工作流标识',
+    name: 'workflowId',
+    helper: '',
+  },
+  workflowNodeId: {
+    label: '工作流节点标识',
+    name: 'workflowNodeId',
+    helper: '',
+  },
+  workflowBranchConditionId: {
+    label: '工作流节点条件标识',
+    name: 'workflowBranchConditionId',
+    helper: '',
+  },
+  name: {
+    label: '名称',
+    name: 'name',
+    helper: '',
+  },
+  description: {
+    label: '简介描述 ',
+    name: 'description',
+    helper: '',
+  },
+  type: {
+    label: '类型',
+    name: 'type',
+    helper: '',
+  },
+  typeNote: {
+    label: '类型',
+    name: 'typeNote',
+    helper: '',
+  },
+  status: {
+    label: '状态',
+    name: 'status',
+    helper: '用户状态',
+  },
+  statusNote: {
+    label: '状态',
+    name: 'statusNote',
+    helper: '用户状态',
+  },
+  createOperatorId: {
+    label: '创建人标识',
+    name: 'createOperatorId',
+    helper: '',
+  },
+  createTime: {
+    label: '创建时间',
+    name: 'createTime',
+    helper: '',
+  },
+  updateOperatorId: {
+    label: '更新人标识',
+    name: 'updateOperatorId',
+    helper: '',
+  },
+  updateTime: {
+    label: '更新时间',
+    name: 'updateTime',
+    helper: '',
+  },
+};