+++ packages/example-remote/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 1.4.87
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-design-playground@1.0.75
+  - antd-management-fast-component@2.6.104
+  - antd-management-fast-flow@1.0.33
+
 ## 1.4.86
 
 ### Patch Changes